Replace Old Plugin System
Replaces the old plugin/loader/msg system with the new routing work
Adds an excerpt to the plugins readme indicating the usage of a 'cmd_args.yaml' file
