+++ common/msg.py
@@ -1,64 +1,114 @@
 
-from collections.abc import MutableMapping
+import collections
+import types
+import uuid
 
-# TODO: Adapt this to be stricter about allowable interface, but looser about allowed interface
-class Message(MutableMapping):
-    def __init__(self, sender):
-        if isinstance(sender, dict):
-            self.msg = sender
+class Message:
+    """
+    Wrap a json dictionary to ensure adherence to the network's message protocol
+    """
+    def __init__(self, plugin=None, role=None):
+        self._msg = {
+            'message_id': uuid.uuid4(),
+            'route': [],
+            'sender': {}
+        }
+        if plugin is not None:
+            self._msg['sender']['uuid'] = plugin.uuid
+        if role is not None:
+            self._msg['sender']['role'] = role
 
-        else:
-            self.msg = {
-                'from': sender,
-                'routing': 'sender',
-                'stop': False
-            }
-
-    def send_to(self, dest):
-        self['routing'] = dest
-
-    def message(self, text):
-        self['text'] = text
-
-    def action(self, key):
-        self['action'] = key
-
-    def dispatch(self, query):
-        self.action('dispatch')
-        self.send_to('dispatch')
-        self['dispatch'] = query
-
-    def finalize(self):
-        return self.msg
+    @property
+    def json_packet(self):
+        return self._msg
 
     @staticmethod
-    def quit():
-        return "quit"
+    def from_json(json_msg):
+        msg = Message()
+        msg._msg = json_msg.copy()
 
-    @staticmethod
-    def stop():
-        return "stop"
+    @property
+    def routing(self):
+        routing_info = {}
+        for key in ['sender', 'route', 'dest', 'forward', 'message_id', 'parent_id', 'ack_uuid']:
+            if key in self._msg:
+                routing_info[key] = self._msg[key].copy()
+        return routing_info
+
+    @property
+    def id(self):
+        return self._msg['message_id']
+
+    @property
+    def parent_id(self):
+        return self._msg.get('parent_id')
+
+    @parent_id.setter
+    def parent_id(self, pid):
+        self._msg['parent_id'] = pid
+
+    @property
+    def action(self):
+        return self._msg.get('action')
+
+    @action.setter
+    def action(self, action):
+        self._msg['action'] = action
+
+    @property
+    def args(self):
+        if 'args' not in self._msg:
+            self._msg['args'] = []
+
+        return self._msg['args']
+
+    @args.setter
+    def args(self, args):
+        if isinstance(args, tuple) or isinstance(args, types.GeneratorType) or isinstance(args, collections.KeysView):
+            args = list(args)
+
+        elif not isinstance(args, list):
+            args = [ args ]
+
+        self._msg['args'] = args
+
+    @property
+    def response(self):
+        return self._msg.get('resp')
+
+    @response.setter
+    def response(self, resp):
+        self._msg['resp'] = resp
+
+
+    def return_to_sender(self):
+        """
+        Immediately sets the destination routing field to be the same as the sender field
+        This has the side-effect of forcing routing to return the message to the sender
+        """
+        self._msg['dest'] = self._msg['sender'].copy()
+
+    def send_to(self, role=None, addr=None, uuid=None, intra_device=None):
+        """
+        Adds routing request information to the destination field of the message
+        NOTE: Depending on the fields used, this will not guarantee sending to a specific app
+        """
+        if 'dest' not in self._msg:
+            self._msg['dest'] = {}
+        if role is not None:
+            self._msg['dest']['role'] = role
+        if addr is not None:
+            self._msg['dest']['addr'] = addr
+        if uuid is not None:
+            self._msg['dest']['uuid'] = uuid
+        if intra_device is not None:
+            self._msg['dest']['intra_device'] = intra_device
 
     @staticmethod
     def is_quit(msg):
-        return msg == Message.quit() or msg == Message.stop()
+        if not isinstance(msg, Message):
+            return True
+        return msg.action == Message.QUIT or msg.action == Message.STOP
 
-
-    # Methods to make the message mimic it's underlying dict typing
-    def __str__(self):
-        return str(self.msg)
-
-    def __delitem__(self, key):
-        del self.msg[key]
-
-    def __getitem__(self, key):
-        return self.msg[key]
-
-    def __iter__(self):
-        return iter(self.msg)
-
-    def __len__(self):
-        return len(self.msg)
-
-    def __setitem__(self, key, value):
-        self.msg[key] = value
+    STOP = "stop"
+    QUIT = "quit"
