+++ docs/todo.txt
@@ -3,9 +3,8 @@
 Short Term:
   Integrate Seshat with device-manager
     Rework the routing system according to 'routing.md'
-      Replace the old loader
-        Rewrite the plugins to work off of the new system
-          Make the wit_ai key a config argument (TODO: Relearn how the plugin system works)
+      Make the wit_ai key a config argument
+        TODO: Need to learn how the config system works
       Rewrite the rust forwarding system to utilize the new routing system
         NOTE: Use hardcoded json paths for now
           Any message crate would likely need to be auto-generated
@@ -26,6 +25,10 @@ Short Term:
     Look at allowing the device-manager to load a config file
       This would better allow us to configure the root folders (in case there are multiple)
         NOTE: We could possibly do this in the command line alone
+    Look at adding Notifications to the communication spec
+      Right now, it's up to the server implementation about how to respond
+        I think, in most cases a response is automatic
+        It would be nice to be able to specify that on the client side
   Add in "device-querying" capabilities to the cli interface
     NOTE: cli-specific interactions to produce statistics about the ai operations
     Come up with a data-collection framework
